Fixed warnings in spell_parser.c

+++ src/spell_parser.c
@@ -109,7 +109,7 @@ void say_spell(struct char_data *ch, int spellnum, struct char_data *tch,
     for (j = 0; *(syls[j].org); j++) {
       if (!strncmp(syls[j].org, lbuf + ofs, strlen(syls[j].org))) {
 	strcat(buf, syls[j].news);	/* strcat: BAD */
-	ofs += strlen(syls[j].org);
+    ofs += (int)strlen(syls[j].org);
         break;
       }
     }
